+++ storage/tianmu/handler/ha_tianmu.cpp
@@ -15,7 +15,7 @@
    Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1335 USA
 */
 
-#include "tianmu_handler.h"
+#include "ha_tianmu.h"
 
 #include <iostream>
